+++ -Wk2_Prac2/spr/Rock.h
@@ -8,7 +8,7 @@ class Rock : public Move {
   Rock(/* args */);
   void setMove(std::string change_);
   std::string getName();
-  int getid();
+  float getid();
   Move* WinOrLoss(Move* compare);
   ~Rock();
 };
