+++ -Wk2_Prac2/spr/Scissors.h
@@ -8,7 +8,7 @@ class Scissors : public Move {
   Scissors(/* args */);
   void setMove(std::string change_);
   std::string getName();
-  int getid();
+  float getid();
   Move* WinOrLoss(Move* compare);
   ~Scissors();
 };
